+++ DSCResources/VE_XD7Features/VE_XD7Features.psm1
@@ -1,10 +1,12 @@
-Import-LocalizedData -BindingVariable localizedData -FileName VE_XD7Features.Resources.psd1;
+Import-LocalizedData -BindingVariable localizedData -FileName VE_XD7Features.Resources.psd1
 
-function Get-TargetResource {
+function Get-TargetResource
+{
     [CmdletBinding()]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSDSCUseVerboseMessageInDSCResource', '')]
     [OutputType([System.Collections.Hashtable])]
-    param (
+    param
+    (
         [Parameter(Mandatory)]
         [ValidateSet('Yes')]
         [System.String] $IsSingleInstance,
@@ -27,27 +29,29 @@ function Get-TargetResource {
         [ValidateSet('Present','Absent')]
         [System.String] $Ensure = 'Present'
     )
-    process {
+    process
+    {
 
         $targetResource = @{
-            IsSingleInstace = $IsSingleInstance;
-            SourcePath = $SourcePath;
-            Role = GetXDInstalledRole -Role $Role;
-            Ensure = 'Absent';
+            IsSingleInstace = $IsSingleInstance
+            SourcePath      = $SourcePath
+            Role            = GetXDInstalledRole -Role $Role
+            Ensure          = 'Absent'
         }
-        if (TestXDInstalledRole -Role $Role) {
-            $targetResource['Ensure'] = 'Present';
+        if (TestXDInstalledRole -Role $Role)
+        {
+            $targetResource['Ensure'] = 'Present'
         }
-        return $targetResource;
+        return $targetResource
+    }
+}
 
-    } #end process
-} #end function Get-TargetResource
-
-
-function Test-TargetResource {
+function Test-TargetResource
+{
     [CmdletBinding()]
     [OutputType([System.Boolean])]
-    param (
+    param
+    (
         [Parameter(Mandatory)]
         [ValidateSet('Yes')]
         [System.String] $IsSingleInstance,
@@ -70,33 +74,37 @@ function Test-TargetResource {
         [ValidateSet('Present','Absent')]
         [System.String] $Ensure = 'Present',
 
-        [Parameter()] [ValidateNotNullOrEmpty()]
-        [System.String] $LogPath = (Join-Path -Path $env:TMP -ChildPath '\Citrix\XenDesktop Installer')
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [System.String] $LogPath = (Join-Path -Path $env:TMP -ChildPath '\Citrix\XenDesktop Installer'),
+
+        [Parameter()]
+        [System.Boolean] $IgnoreHardwareCheckFailure
     )
     process {
 
-        $targetResource = Get-TargetResource @PSBoundParameters;
-        if ($Ensure -eq $targetResource.Ensure) {
-
-            Write-Verbose ($localizedData.ResourceInDesiredState -f ($Role -join ','));
-            return $true;
+        $targetResource = Get-TargetResource -IsSingleInstance $IsSingleInstance -SourcePath $SourcePath -Role $Role
+        if ($Ensure -eq $targetResource.Ensure)
+        {
+            Write-Verbose ($localizedData.ResourceInDesiredState -f ($Role -join ','))
+            return $true
         }
-        else {
-
-            Write-Verbose ($localizedData.ResourceNotInDesiredState -f ($Role -join ','));
-            return $false;
+        else
+        {
+            Write-Verbose ($localizedData.ResourceNotInDesiredState -f ($Role -join ','))
+            return $false
         }
+    }
+}
 
-    } #end process
-} #end function Test-TargetResource
-
-
-function Set-TargetResource {
+function Set-TargetResource
+{
     [CmdletBinding()]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions', '')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseDeclaredVarsMoreThanAssignments', 'DSCMachineStatus')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidGlobalVars', 'global:DSCMachineStatus')]
-    param (
+    param
+    (
         [Parameter(Mandatory)]
         [ValidateSet('Yes')]
         [System.String] $IsSingleInstance,
@@ -119,62 +127,77 @@ function Set-TargetResource {
         [ValidateSet('Present','Absent')]
         [System.String] $Ensure = 'Present',
 
-        [Parameter()] [ValidateNotNullOrEmpty()]
-        [System.String] $LogPath = (Join-Path -Path $env:TMP -ChildPath '\Citrix\XenDesktop Installer')
-    )
-    begin {
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [System.String] $LogPath = (Join-Path -Path $env:TMP -ChildPath '\Citrix\XenDesktop Installer'),
 
-        if (-not (Test-Path -Path $SourcePath -PathType Container)) {
+        [Parameter()]
+        [System.Boolean] $IgnoreHardwareCheckFailure
+    )
+    begin
+    {
+        if (-not (Test-Path -Path $SourcePath -PathType Container))
+        {
             throw ($localizedData.InvalidSourcePathError -f $SourcePath);
         }
-
     }
-    process {
-
-        if ($Ensure -eq 'Present') {
-
-            Write-Verbose ($localizedData.InstallingRole -f ($Role -join ','));
-            $installArguments = ResolveXDServerSetupArguments -Role $Role -LogPath $LogPath;
+    process
+    {
+        if ($Ensure -eq 'Present')
+        {
+            Write-Verbose ($localizedData.InstallingRole -f ($Role -join ','))
+            $resolveXDServerSetupArgumentsParams = @{
+                Role = $Role
+                LogPath = $LogPath
+                IgnoreHardwareCheckFailure = $IgnoreHardwareCheckFailure
+            }
+            $installArguments = ResolveXDServerSetupArguments @resolveXDServerSetupArgumentsParams
         }
-        else {
-
+        else
+        {
             ## Uninstall
-            Write-Verbose ($localizedData.UninstallingRole -f ($Role -join ','));
-            $installArguments = ResolveXDServerSetupArguments -Role $Role -LogPath $LogPath -Uninstall;
+            Write-Verbose ($localizedData.UninstallingRole -f ($Role -join ','))
+            $resolveXDServerSetupArgumentsParams = @{
+                Role = $Role
+                LogPath = $LogPath
+                Uninstall = $true
+            }
+            $installArguments = ResolveXDServerSetupArguments @resolveXDServerSetupArgumentsParams
         }
 
-        Write-Verbose ($localizedData.LogDirectorySet -f $logPath);
-        Write-Verbose ($localizedData.SourceDirectorySet -f $SourcePath);
+        Write-Verbose ($localizedData.LogDirectorySet -f $logPath)
+        Write-Verbose ($localizedData.SourceDirectorySet -f $SourcePath)
 
         $startWaitProcessParams = @{
-            FilePath = ResolveXDSetupMedia -Role $Role -SourcePath $SourcePath;
-            ArgumentList = $installArguments;
+            FilePath = ResolveXDSetupMedia -Role $Role -SourcePath $SourcePath
+            ArgumentList = $installArguments
         }
 
-        if ($PSBoundParameters.ContainsKey('Credential')) {
-            $startWaitProcessParams['Credential'] = $Credential;
+        if ($PSBoundParameters.ContainsKey('Credential'))
+        {
+            $startWaitProcessParams['Credential'] = $Credential
         }
 
-        $exitCode = StartWaitProcess @startWaitProcessParams -Verbose:$Verbose;
+        $exitCode = StartWaitProcess @startWaitProcessParams -Verbose:$Verbose
         # Check for reboot
-        if (($exitCode -eq 3010) -or ($Role -contains 'Controller')) {
-            $global:DSCMachineStatus = 1;
+        if (($exitCode -eq 3010) -or ($Role -contains 'Controller'))
+        {
+            $global:DSCMachineStatus = 1
         }
-        elseif ($Role -contains 'Storefront') {
+        elseif ($Role -contains 'Storefront')
+        {
             ## Add the Storefront module path to the current session
-            if (Test-Path -Path "$env:ProgramFiles\Citrix\Receiver StoreFront\PowerShellSDK\Modules\") {
-                $env:PSModulePath += ";$env:ProgramFiles\Citrix\Receiver StoreFront\PowerShellSDK\Modules\";
+            if (Test-Path -Path "$env:ProgramFiles\Citrix\Receiver StoreFront\PowerShellSDK\Modules\")
+            {
+                $env:PSModulePath += ";$env:ProgramFiles\Citrix\Receiver StoreFront\PowerShellSDK\Modules\"
             }
         }
-
-    } #end process
-} #end function Set-TargetResource
-
-
-$moduleRoot = Split-Path -Path $MyInvocation.MyCommand.Path -Parent;
+    }
+}
 
 ## Import the XD7Common library functions
-$moduleParent = Split-Path -Path $moduleRoot -Parent;
-Import-Module (Join-Path -Path $moduleParent -ChildPath 'VE_XD7Common');
+$moduleRoot = Split-Path -Path $MyInvocation.MyCommand.Path -Parent
+$moduleParent = Split-Path -Path $moduleRoot -Parent
+Import-Module (Join-Path -Path $moduleParent -ChildPath 'VE_XD7Common')
 
-Export-ModuleMember -Function *-TargetResource;
+Export-ModuleMember -Function *-TargetResource